requirements.txt
# Flask core
Flask==3.0.0
Werkzeug==3.0.1

# Base de datos
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
alembic==1.13.1

# Autenticación y seguridad
PyJWT==2.8.0
Flask-WTF==1.2.1

# Serialización y validación
marshmallow==3.20.1

# Documentación API
flasgger==0.9.7.1

# Caché
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Servidor WSGI
gunicorn==21.2.0

# Utilidades
python-dotenv==1.0.0